Separate the template modal dialog as an individual component.
