--- webportal/src/app/marketplace/template-view/resource-table.component.ejs
@@ -1,1 +1,0 @@
-<table id="resource-table" class="table table-bordered table-hover table-condensed"></table>
