+++ webportal/src/app/marketplace/template-view/template-modal.component.ejs
@@ -1,0 +1,28 @@
+<div class="modal fade" id="shareModal" tabindex="-1" role="dialog" aria-labelledby="modalLable" aria-hidden="true">
+  <div class="modal-dialog" role="document">
+    <div class="modal-content">
+      <div class="modal-header">
+        <h4 class="modal-title" id="modalLable">
+          Share My Template
+        </h4>
+      </div>
+      <div class="modal-body">
+        <div class="form-group">
+          <input class="btn btn-default" id="file-template" type="file" accept=".yaml">
+          <p class="help-block">The file path of the job template to submit.</p>
+        </div>
+        <div class="form-group" id="form-table">
+          <table id="resource-table" class="table table-bordered table-hover table-condensed"></table>
+        </div>
+      </div>
+      <div class="modal-footer">
+        <button type="button" class="btn btn-primary" id="btn-submit">
+          Submit
+        </button>
+        <button type="button" class="btn btn-default" data-dismiss="modal">
+          Close
+        </button>
+      </div>
+    </div>
+  </div>
+</div>
