+++ webportal/src/app/marketplace/template-view/template-view.component.js
@@ -23,25 +23,21 @@ require('datatables.net-bs/css/dataTables.bootstrap.css');
 require('datatables.net-plugins/sorting/natural.js');
 require('./template-view.component.scss');
 
-const yaml = require('js-yaml');
 const breadcrumbComponent = require('../../job/breadcrumb/breadcrumb.component.ejs');
 const loadingComponent = require('../../job/loading/loading.component.ejs');
 const templateViewComponent = require('./template-view.component.ejs');
 const templateTableComponent = require('./template-table.component.ejs');
-const resourceTableComponent = require('./resource-table.component.ejs');
+const templateModalComponent = require('./template-modal.component');
 const loading = require('../../job/loading/loading.component');
 const webportalConfig = require('../../config/webportal.config.json');
-const userAuth = require('../../user/user-auth/user-auth.component');
 
 let templateTable = null;
-let resourceTable = null;
 
-const templateViewHtml = templateViewComponent({
+$('#content-wrapper').html(templateViewComponent({
   breadcrumb: breadcrumbComponent,
   loading: loadingComponent,
-  templateTable: templateTableComponent,
-  resourceTable: resourceTableComponent
-});
+  templateTable: templateTableComponent
+}));
 
 const generateQueryString = function(data) {
   return '?name=' + encodeURIComponent(data.name) + '&version='
@@ -101,8 +97,8 @@ const loadTemplates = function() {
         data: 'name'
       },
       {
-        title: 'Type',
-        data: 'type'
+        title: 'Version',
+        data: 'version'
       },
       {
         title: 'Constributor',
@@ -152,141 +148,24 @@ const loadTemplates = function() {
   }).api();
 };
 
-const resizeContentWrapper = function(event) {
+$(window).resize(function(event) {
   $('#content-wrapper').css({'height': $(window).height() + 'px'});
   if (templateTable != null) {
-    $('.dataTables_scrollBody').css('height', (($(window).height() - 265)) + 'px');
+    $('.dataTables_scrollBody').css('height', (($(window).height() - 315)) + 'px');
     templateTable.columns.adjust().draw();
   }
-};
+});
 
-const analyzeFile = function(source) {
-  $('#form-table').html(resourceTableComponent());
-  resourceTable = null;
-  if (source.files && source.files[0]) {
-    if (window.FileReader) {
-      var file = source.files[0];
-      var fr = new FileReader();
-      fr.onload = function(e) {
-        if (e.target.result) {
-          try {
-            var data = yaml.safeLoad(e.target.result);
-            resources = [
-              {
-                'name': data.job.name,
-                'type': 'job',
-                'version': data.job.version
-              }
-            ];
-            data.prerequisites.forEach(function (element) {
-              resources.push({
-                'name': element.name,
-                'type': element.type,
-                'version': element.version
-              });
-            });
-            resourceTable = $('#resource-table').DataTable({
-              data: resources,
-              columns: [
-                {
-                  title: 'Name',
-                  data: 'name'
-                },
-                {
-                  title: 'Type',
-                  data: 'type'
-                },
-                {
-                  title: 'Version',
-                  data: 'version'
-                },
-                {
-                  title: 'Include',
-                  data: null,
-                  orderable: false,
-                  searchable: false,
-                  render: function(data, type) {
-                    return `<input type="checkbox" name="included" value="${data.name}:${data.version}" checked="checked" />`;
-                  }
-                }
-              ],
-              'order': [
-                [0, 'asc']
-              ],
-              'autoWidth': false,
-              'deferRender': true,
-              'paging': false,
-              'info': false,
-              'searching': false
-            });
-            resourceTable.originData = data;
-            resourceTable.originFileName = file.name;
-            return;
-          } catch (e) {
-            if (e.message) {
-              return alert(e);
-            }
-          }
-        }
-        alert('Failed to read the selected file.');
-      };
-      fr.readAsText(file);
-    } else {
-      alert('The browser does not support preview text file!');
-    }
-  }
-};
-
-const submitTemplate = function(source) {
-  var ajaxData = {
-    'template': resourceTable.originData,
-    'included': [],
-    'filename': resourceTable.originFileName
-  }
-  $('[name="included"]').each(function(index, element) {
-    if (element.checked) {
-      ajaxData['included'].push(element.value);
-    }
-  });
-
-  //$('#shareModal').modal('hide');
-  if (resourceTable != null) {
-    //loading.showLoading();
-    userAuth.checkToken((token) => {
-      $.ajax({
-        type: "POST",
-        url: `${webportalConfig.restServerUri}/api/v1/template`,
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-        data: ajaxData,
-        dataType: 'json',
-        success: function() {
-          alert('success!');
-          //loading.hideLoading();
-        },
-        error: function(xhr, status, error) {
-          var res = JSON.parse(xhr.responseText);
-          alert(res.message ? res.message : res.toString());
-          //loading.hideLoading();
-        }
-      });
-    });
-  }
-};
-
-window.analyzeFile = analyzeFile;
-window.onresize = resizeContentWrapper;
-window.submitTemplate = submitTemplate;
-
-$('#content-wrapper').html(templateViewHtml);
+$('#btn-share').click(function(event) {
+  $('#modalPlaceHolder').html(templateModalComponent.generateHtml());
+  templateModalComponent.initializeComponent();
+  $('#shareModal').modal('show');
+});
 
 $(document).ready(() => {
   $('#sidebar-menu--template-view').addClass('active');
   $('#content-wrapper').css({'overflow': 'hidden'});
-  resizeContentWrapper();
+  window.dispatchEvent(new Event('resize'));
   loadTemplates();
   loadRecommended();
 });
-
-module.exports = {loadTemplates};
